requirements.txt
attrs==20.2.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
zipp==3.4.0
